image_analysis/requirements.txt
torch>=1.13.1
torchvision>=0.14.1
Pillow>=9.0.0
numpy>=1.21.0
ftfy
regex
tqdm
clip @ git+https://github.com/openai/CLIP.git